chore: Breakup large spec file to avoid chrome crash
